+++ database/migrations/2024_06_05_144000_ragnarok_fara_allow_negative_sales_amount.php
@@ -1,0 +1,29 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+/**
+ * Allow negative value for sales amount.
+ */
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('fara_stat_traffic_income', function (Blueprint $table) {
+            $table->float('amountofproductssale', 10, 3)->nullable()->default(NULL)->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        //
+    }
+};